Put the classes for logistic regression in one file
