+++ ethicml/algorithms/inprocess/__init__.py
@@ -4,9 +4,7 @@ from .in_algorithm import InAlgorithm, InAlgorithmAsync
 from .fair_gpyt import GPyT, GPyTDemPar, GPyTEqOdds
 from .kamishima import Kamishima
 from .kamiran import Kamiran
-from .logistic_regression import LR
-from .logistic_regression_cross_validated import LRCV
-from .logistic_regression_probability import LRProb
+from .logistic_regression import LR, LRCV, LRProb
 from .svm import SVM
 from .majority import Majority
 from .mlp import MLP
